Reorganize the code (move implementations and abstract methods to the end of the classes), move PlayerDataContainer into a new Containers package
