+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.redcreator37</groupId>
     <artifactId>PlayerManagement</artifactId>
-    <version>1.6</version>
+    <version>1.7</version>
     <packaging>jar</packaging>
 
     <name>PlayerManagement</name>
