+++ src/main/java/com/redcreator37/playermanagement/PreferencesHolder.java
@@ -22,119 +22,84 @@ public class PreferencesHolder {
      * Any in-game console output will get prefixed by this
      */
     public String prefix = getDefaultPrefix();
-
-    /**
-     * The database to use to store the player data
-     */
-    String databasePath = "PlayerData.db";
-
     /**
      * The date format to be used, <strong>must</strong>
      * be a valid standard date format
      */
     public String dateFormat = "yyyy-MM-dd";
-
     /**
      * The standard price of an ID card that's charged when a player
      * loses theirs and requests a new one
      */
     public double cardPrice = 2500;
-
     /**
      * The item lore (description) text which separates ID cards from
      * other items
      */
     public String cardItemLore = "Unique Player ID Card";
-
     /**
      * Enables experimental features
      */
     public boolean experimentalFeatures = false;
-
     /**
      * The price of establishing a new company
      */
     public double establishPrice = 3000;
-
     /**
      * The amount of money that'll be taken after a player is fined
      */
     public double fineAmount = 3000;
-
     /**
      * The number of fines a player can receive before they get banned
      */
     public int maxFines = 5;
-
     /**
      * The text which will be displayed on the top of each scoreboard
      * sign
      */
     public String scoreboardSignText = "§8[§9TopPlayers§8]";
-
     /**
      * Toggles automatic economy management
      */
     public boolean autoEcoEnabled = true;
-
     /**
      * The amount of times between the runs of the automatic economy
      * background task
      */
     public int autoEcoTimeSeconds = 1200;
-
     /**
      * The minimal wage (basic income), paid to unemployed players
      */
     public EconomyProvider.MinimalWage minimalWage =
             new EconomyProvider.MinimalWage(1000, 250);
-
     /**
      * Toggles advancement rewards
      */
     public boolean rewardsEnabled = true;
-
     /**
      * The reward for rank 1 advancements
      */
     public double rewardRank1 = 10;
-
     /**
      * The reward for rank 2 advancements
      */
     public double rewardRank2 = 25;
-
     /**
      * The reward for rank 3 advancements
      */
     public double rewardRank3 = 50;
-
     /**
      * The list of rank 1 advancements
      */
     public List<String> advRank1 = AdvancementReward.getAdvancements(1);
-
     /**
      * The list of rank 2 advancements
      */
     public List<String> advRank2 = AdvancementReward.getAdvancements(2);
-
     /**
      * The list of rank 3 advancements
      */
     public List<String> advRank3 = AdvancementReward.getAdvancements(3);
-
-    /**
-     * Toggles the enhanced TAB player list
-     */
-    boolean playerListEnabled = true;
-
-    /**
-     * The amount of time in seconds between player list updates.
-     * This value is ignored when a player joins / leaves the server.
-     */
-    int playerListUpdateSeconds = 300;
-
     /**
      * Player list header, displayed above the list of players
      */
@@ -142,7 +107,6 @@ public class PreferencesHolder {
             "§b§oA Minecraft Server                                   §r§a{playercount}§8/§a{maxplayers}",
             "§8--------------------------------------------"
     };
-
     /**
      * Player list footer, displayed below the list of players
      */
@@ -157,47 +121,62 @@ public class PreferencesHolder {
             "§4YouTube:   §r§f§nlink§r      §r§8: :      §9Discord:   §r§f§nlink§r             ",
             "\n§8§oPowered by PlayerManagement v1.6 by RedCreator37"
     };
-
     /**
      * The style of a player entry in the player list for players
      * without ranks
      */
     public String genericPlayerEntry = "§8[ §f{playername} §8]";
-
     /**
      * The style of a player entry in the player list for members
      */
     public String memberPlayerEntry = "§8[ §a{playername} §8]";
-
     /**
      * The style of a player entry in the player list for VIP players
      */
     public String vipPlayerEntry = "§8[ §b{playername} §8]";
-
     /**
      * The style of a player entry in the player list for admins
      */
     public String adminPlayerEntry = "§8[ §c{playername} §8]";
-
     /**
      * The rank label for players without ranks
      */
     public String genericPlayerLabel = "§fNone";
-
     /**
      * The rank label for members
      */
     public String memberPlayerLabel = "§aMembers";
-
     /**
      * The rank label for VIP players
      */
     public String vipPlayerLabel = "§bVIP";
-
     /**
      * The rank label for admins
      */
     public String adminPlayerLabel = "§cAdmins";
+    /**
+     * The database to use to store the player data
+     */
+    String databasePath = "PlayerData.db";
+    /**
+     * Toggles the enhanced TAB player list
+     */
+    boolean playerListEnabled = true;
+    /**
+     * The amount of time in seconds between player list updates.
+     * This value is ignored when a player joins / leaves the server.
+     */
+    int playerListUpdateSeconds = 300;
+
+    /**
+     * Returns the plugin's localized default chat output prefix
+     *
+     * @return the formatted string
+     */
+    private static String getDefaultPrefix() {
+        return ChatColor.DARK_GRAY + "[" + ChatColor.BLUE
+                + Localization.lc("server") + ChatColor.DARK_GRAY + "] ";
+    }
 
     /**
      * Loads the values from the config file or generates them if they
@@ -292,16 +271,6 @@ public class PreferencesHolder {
         memberPlayerLabel = conf.getString("PlayerList.Label.Members");
         vipPlayerLabel = conf.getString("PlayerList.Label.VIP");
         adminPlayerLabel = conf.getString("PlayerList.Label.Admins");
-    }
-
-    /**
-     * Returns the plugin's localized default chat output prefix
-     *
-     * @return the formatted string
-     */
-    private static String getDefaultPrefix() {
-        return ChatColor.DARK_GRAY + "[" + ChatColor.BLUE
-                + Localization.lc("server") + ChatColor.DARK_GRAY + "] ";
     }
 
 }
