+++ src/main/java/com/redcreator37/playermanagement/TopPlayerList.java
@@ -20,29 +20,6 @@ import static com.redcreator37.playermanagement.PlayerManagement.eco;
 public class TopPlayerList implements Listener {
 
     /**
-     * Handles the sign text change event
-     *
-     * @param event the sign change event
-     */
-    @EventHandler(priority = EventPriority.HIGH)
-    public void eventSignChanged(SignChangeEvent event) {
-        if (!event.getLines()[0].equalsIgnoreCase("[topplayers]")) return;
-
-        event.setLine(0, PlayerManagement.prefs.scoreboardSignText);
-        List<String> topPlayers = getTopPlayers(Arrays
-                .asList(Bukkit.getOfflinePlayers())).stream()
-                .distinct().collect(Collectors.toList());
-
-        for (int i = 0, size = topPlayers.size(); i < size; i++) {
-            String p = topPlayers.get(i);
-            String text = "§4" + (i + 1) + ".§r ";
-            text += p != null ? p : "§oN/A";
-            if (i > 3) break;   // in-game sign entity constraint
-            event.setLine(i + 1, text);
-        }
-    }
-
-    /**
      * Returns a sorted list of usernames of top players
      * from this list
      *
@@ -64,6 +41,29 @@ public class TopPlayerList implements Listener {
         players.sort((lhs, rhs) -> Double
                 .compare(eco.getBalance(rhs), eco.getBalance(lhs)));
         return players;
+    }
+
+    /**
+     * Handles the sign text change event
+     *
+     * @param event the sign change event
+     */
+    @EventHandler(priority = EventPriority.HIGH)
+    public void eventSignChanged(SignChangeEvent event) {
+        if (!event.getLines()[0].equalsIgnoreCase("[topplayers]")) return;
+
+        event.setLine(0, PlayerManagement.prefs.scoreboardSignText);
+        List<String> topPlayers = getTopPlayers(Arrays
+                .asList(Bukkit.getOfflinePlayers())).stream()
+                .distinct().collect(Collectors.toList());
+
+        for (int i = 0, size = topPlayers.size(); i < size; i++) {
+            String p = topPlayers.get(i);
+            String text = "§4" + (i + 1) + ".§r ";
+            text += p != null ? p : "§oN/A";
+            if (i > 3) break;   // in-game sign entity constraint
+            event.setLine(i + 1, text);
+        }
     }
 
 }
